requirements.txt
# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.4.0

# Feature store
feast==0.36.0

# Redis
redis==5.0.1

# ML
xgboost==2.0.3
scikit-learn==1.4.0
imbalanced-learn==0.12.0
numpy==1.26.4
pandas==2.2.0

# MLflow
mlflow==2.11.1

# Serving
bentoml==1.2.2
fastapi==0.110.0
uvicorn==0.27.1

# Monitoring
evidently==0.4.22
prometheus-client==0.20.0

# Orchestration
apache-airflow==2.8.2

# Utilities
python-dotenv==1.0.1
pyarrow==15.0.0
boto3==1.34.0